requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
fastapi-mail==1.4.1
authlib==1.2.1
httpx==0.25.2
pyotp==2.9.0
qrcode[pil]==7.4.2
python-dotenv==1.0.0
redis==5.0.1
user-agents==2.2.0